refactor: CategoryStackView 값이 없을때 기본값추가, AlertManager 싱글톤 객체 추가
