+++ BoxOffice/ViewController/BoxOfficeViewController.swift
@@ -90,23 +90,11 @@ final class BoxOfficeViewController: UIViewController {
             case .failure:
                 DispatchQueue.main.async {
                     self?.activityIndicator.stopAnimating()
-                    self?.showAlert()
+                    let alertController = AlertManager.shared.showFailureAlert()
+                    self?.present(alertController, animated: true)
                 }
             }
         }
-    }
-    
-    private func showAlert() {
-        let alert = UIAlertController(
-            title: "에러",
-            message: "데이터를 불러올 수 없습니다",
-            preferredStyle: .alert
-        )
-        let okAction = UIAlertAction(title: "확인", style: .default)
-        
-        alert.addAction(okAction)
-        
-        present(alert, animated: true)
     }
     
     @objc private func refresh() {
@@ -134,7 +122,8 @@ final class BoxOfficeViewController: UIViewController {
                 }
             case .failure:
                 DispatchQueue.main.async {
-                    self?.showAlert()
+                    let alertController = AlertManager.shared.showFailureAlert()
+                    self?.present(alertController, animated: true)
                     self?.refreshControl.endRefreshing()
                 }
             }
